Merge changes from Bitmessage/PyBitmessage

+++ src/singleton.py
@@ -1,25 +1,44 @@
 #! /usr/bin/env python
 
-import sys
-import os
+import atexit
 import errno
-import tempfile
 from multiprocessing import Process
+import os
+import sys
+import shared
 
+try:
+    import fcntl  # @UnresolvedImport
+except:
+    pass
 
 class singleinstance:
     """
-    Implements a single instance application by creating a lock file based on the full path to the script file.
+    Implements a single instance application by creating a lock file at appdata.
 
     This is based upon the singleton class from tendo https://github.com/pycontribs/tendo
     which is under the Python Software Foundation License version 2    
     """
-    def __init__(self, flavor_id=""):
-        import sys
+    def __init__(self, flavor_id="", daemon=False):
         self.initialized = False
-        basename = os.path.splitext(os.path.abspath(sys.argv[0]))[0].replace("/", "-").replace(":", "").replace("\\", "-") + '-%s' % flavor_id + '.lock'
-        self.lockfile = os.path.normpath(tempfile.gettempdir() + '/' + basename)
+        self.counter = 0
+        self.daemon = daemon
+        self.lockPid = None
+        self.lockfile = os.path.normpath(os.path.join(shared.appdata, 'singleton%s.lock' % flavor_id))
 
+        if not self.daemon and not shared.curses:
+            # Tells the already running (if any) application to get focus.
+            import bitmessageqt
+            bitmessageqt.init()
+
+        self.lock()
+
+        self.initialized = True
+        atexit.register(self.cleanup)
+
+    def lock(self):
+        if self.lockPid is None:
+            self.lockPid = os.getpid()
         if sys.platform == 'win32':
             try:
                 # file already exists, we try to remove (in case previous execution was interrupted)
@@ -34,28 +53,32 @@ class singleinstance:
                 print(e.errno)
                 raise
         else:  # non Windows
-            import fcntl  # @UnresolvedImport
             self.fp = open(self.lockfile, 'w')
             try:
-                fcntl.lockf(self.fp, fcntl.LOCK_EX | fcntl.LOCK_NB)
+                if self.daemon and self.lockPid != os.getpid():
+                    fcntl.lockf(self.fp, fcntl.LOCK_EX) # wait for parent to finish
+                else:
+                    fcntl.lockf(self.fp, fcntl.LOCK_EX | fcntl.LOCK_NB)
+                self.lockPid = os.getpid()
             except IOError:
                 print 'Another instance of this application is already running'
                 sys.exit(-1)
-        self.initialized = True
 
-    def __del__(self):
-        import sys
+    def cleanup(self):
         if not self.initialized:
             return
+        if self.daemon and self.lockPid == os.getpid():
+            # these are the two initial forks while daemonizing
+            return
+        print "Cleaning up lockfile"
         try:
             if sys.platform == 'win32':
                 if hasattr(self, 'fd'):
                     os.close(self.fd)
                     os.unlink(self.lockfile)
             else:
-                import fcntl  # @UnresolvedImport
                 fcntl.lockf(self.fp, fcntl.LOCK_UN)
                 if os.path.isfile(self.lockfile):
                     os.unlink(self.lockfile)
         except Exception, e:
-            sys.exit(-1)
+            pass
